+++ packages/twenty-server/src/database/commands/upgrade-version-command/upgrade-version-command.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { WorkspaceIteratorModule } from 'src/database/commands/command-runners/workspace-iterator.module';
-import { CoreMigrationRunnerModule } from 'src/database/commands/core-migration-runner/core-migration-runner.module';
+import { CoreMigrationModule } from 'src/database/commands/core-migration/core-migration.module';
 import { V1_20_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-version-command/1-20/1-20-upgrade-version-command.module';
 import { V1_21_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-version-command/1-21/1-21-upgrade-version-command.module';
 import { UpgradeCommand } from 'src/database/commands/upgrade-version-command/upgrade.command';
@@ -18,7 +18,7 @@ import { WorkspaceVersionModule } from 'src/engine/workspace-manager/workspace-v
     V1_21_UpgradeVersionCommandModule,
     DataSourceModule,
     CoreEngineVersionModule,
-    CoreMigrationRunnerModule,
+    CoreMigrationModule,
     WorkspaceVersionModule,
     WorkspaceIteratorModule,
   ],
